refactor: model/ 을 도메인 폴더 하위로 변경
